Revert "Fix lint, move to adapters: PlayerStatsRecyclerViewAdapter"
This reverts commit e09b8eded75cbc99d5c1ee1ed147928f5f7f9fcf.
